Main part refactored and a module added.
Moved gpx realted, pause reading and track reading parts outside as a module named 'nst.py'.
